+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version [0.21.0](https://github.com/cedx/base/compare/v0.20.0...v0.21.0)
+- Added the `show()` and `hide()` methods to the `UI.Components.OfflineIndicator` component. 
 - Renamed the `Animation` property of UI components to `Fade`.
 - Replaced the `Hidden` property of UI components by the `Open` property.